Implement variance and use floats in LSE

+++ firmware/inc/data_processor.h
@@ -15,14 +15,14 @@ struct Corner {
 
 struct LSE_data {
   uint8_t N;
-  int16_t u_sum;
-  int16_t v_sum;
-  int32_t rx_sum;
-  int32_t ry_sum;
-  uint32_t rx2_sum;
-  uint32_t ry2_sum;
-  int32_t rxv_sum;
-  int32_t ryu_sum;
+  float u_sum;
+  float v_sum;
+  float rx_sum;
+  float ry_sum;
+  float rx2_sum;
+  float ry2_sum;
+  float rxv_sum;
+  float ryu_sum;
 };
 
 struct LSE_solution {
